Funcionalidad del CRUD de Vehiculos + documento de manual de usuario

+++ TrabajoPracticoIntegrador/Entity/Vehiculo.cs
@@ -7,10 +7,10 @@
         public string Modelo { get; set; }
         public int Anio { get; set; }
         public int Kilometraje { get; set; }
-        public decimal Precio { get; set; }
+        public double Precio { get; set; }
         public string Patente { get; set; }
         public Vehiculo() { }
-        public Vehiculo(string marca, string modelo, int anio, int kilometraje, decimal precio, string patente)
+        public Vehiculo(string marca, string modelo, int anio, int kilometraje, double precio, string patente)
         {
             Marca = marca;
             Modelo = modelo;
